Move launchApp to apps.ts
